Dados dashboard

+++ src/main/java/com/fcfadmin/fcf/controller/FinanceiroController/FinanceiroController.java
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 @RestController @RequestMapping("/api/financeiro") public class FinanceiroController {
 
     private final String STATUS_ABERTO = "Aberto";
-//    private final String STATUS_QUITADO = "Quitado";
+    //    private final String STATUS_QUITADO = "Quitado";
 
     private final String TIPO_RECEBER = "R";
     private final String TIPO_PAGAR = "P";
@@ -24,57 +24,52 @@ import java.util.stream.Collectors;
         this.financeiroRepository = financeiroRepository;
     }
 
-    @RequestMapping("/listreceber")
-    public List<Financeiro> buscarListaReceber() {
+    @RequestMapping("/listreceber") public List<Financeiro> buscarListaReceber() {
         return financeiroRepository.buscarListaReceber();
     }
 
-    @RequestMapping("/listpagar")
-    public List<Financeiro> buscarListaPagar() {
+    @RequestMapping("/listpagar") public List<Financeiro> buscarListaPagar() {
         return financeiroRepository.buscarListaPagar();
     }
 
-    @RequestMapping("/financeiro/{id}")
-    public Financeiro buscarPorid(@PathVariable Long id) {
+    @RequestMapping("/financeiro/{id}") public Financeiro buscarPorid(@PathVariable Long id) {
         return financeiroRepository.findById(id).orElse(null);
     }
 
-    @RequestMapping(value = "/salvar", method = RequestMethod.POST)
-    public Financeiro salvar(@RequestBody Financeiro financeiro) {
+    @RequestMapping(value = "/salvar", method = RequestMethod.POST) public Financeiro salvar(@RequestBody Financeiro financeiro) {
         return financeiroRepository.save(financeiro);
     }
 
-    @RequestMapping(value = "/deletar", method = RequestMethod.POST)
-    public void deletar(@RequestBody Financeiro financeiro) {
+    @RequestMapping(value = "/deletar", method = RequestMethod.POST) public void deletar(@RequestBody Financeiro financeiro) throws Exception {
+
+        Financeiro f = financeiroRepository.findById(financeiro.getId()).get();
+        if (f.getIdContrato() != null && f.getIdContrato() > 0) {
+            throw new Exception("Financeiro não pode ser exluido pois esta vinculado a um contrato");
+        }
         financeiroRepository.delete(financeiro);
     }
 
-    @RequestMapping("/receber-vencido")
-    public List<Financeiro> buscarReceberVencido() {
+    @RequestMapping("/receber-vencido") public List<Financeiro> buscarReceberVencido() {
         List<Financeiro> documentosReceberAberto = financeiroRepository.buscarDadosFinanceiro(TIPO_RECEBER, STATUS_ABERTO);
         return documentosReceberAberto.stream().filter(dado -> dado.getVencimento().compareTo(LocalDate.now()) < 0).collect(Collectors.toList());
     }
 
-    @RequestMapping("/receber-vencendo")
-    public List<Financeiro> buscarReceberVencendo() {
+    @RequestMapping("/receber-vencendo") public List<Financeiro> buscarReceberVencendo() {
         List<Financeiro> documentosReceberAberto = financeiroRepository.buscarDadosFinanceiro(TIPO_RECEBER, STATUS_ABERTO);
         return documentosReceberAberto.stream().filter(dado -> dado.getVencimento().compareTo(LocalDate.now()) == 0).collect(Collectors.toList());
     }
 
-    @RequestMapping("/pagar-vencido")
-    public List<Financeiro> buscarPagarVencido() {
+    @RequestMapping("/pagar-vencido") public List<Financeiro> buscarPagarVencido() {
         List<Financeiro> documentosPagarAberto = financeiroRepository.buscarDadosFinanceiro(TIPO_PAGAR, STATUS_ABERTO);
         return documentosPagarAberto.stream().filter(dado -> dado.getVencimento().compareTo(LocalDate.now()) < 0).collect(Collectors.toList());
     }
 
-    @RequestMapping("/pagar-vencendo")
-    public List<Financeiro> buscarPagarVencendo() {
+    @RequestMapping("/pagar-vencendo") public List<Financeiro> buscarPagarVencendo() {
         List<Financeiro> documentosPagarAberto = financeiroRepository.buscarDadosFinanceiro(TIPO_PAGAR, STATUS_ABERTO);
         return documentosPagarAberto.stream().filter(dado -> dado.getVencimento().compareTo(LocalDate.now()) == 0).collect(Collectors.toList());
     }
 
-    @RequestMapping("/dashboard")
-    public DadosDashboard buscarDadosFinanceiro() {
+    @RequestMapping("/dashboard") public DadosDashboard buscarDadosFinanceiro() {
 
         BigDecimal receberVencido = buscarReceberVencido().stream().map(Financeiro::getValor).reduce(BigDecimal.ZERO, BigDecimal::add);
 
@@ -92,4 +87,6 @@ import java.util.stream.Collectors;
 
         return dashboard;
     }
+
+
 }
